Base support

+++ api_test/base.py
@@ -9,6 +9,7 @@ class BaseAPICase(TestCase):
     session: requests.Session = None
     post_payload: dict = None
     object_id_key = "id"
+    delete_on_teardown = True
 
     @property
     def object_id(self):
@@ -24,11 +25,9 @@ class BaseAPICase(TestCase):
         self.assertTrue(response.ok)
 
     def setUp(self):
-        assert self.base_url
-        assert self.route
-        assert self.post_payload
+        if not self.route:
+            return
 
-        super(BaseAPICase, self).setUp()
         if not self.session:
             self.session = requests.Session()
 
@@ -37,5 +36,6 @@ class BaseAPICase(TestCase):
         self.object = res_obj
 
     def tearDown(self):
-        super(BaseAPICase, self).tearDown()
-        self.delete()
+        super().tearDown()
+        if self.delete_on_teardown and self.route:
+            self.delete()
